refactor: 반복되는 handleLoad를 커스텀훅으로 대체
