+++ src/pages/PostPage/PostPage.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useCallback } from "react";
+import React, { useState, useEffect } from "react";
 import { useParams } from "react-router-dom";
 import styles from "./PostPage.module.css";
 import logoImage from "../../assets/images/logo.png";
@@ -6,38 +6,13 @@ import PostProfile from "../../components/PostProfile/PostProfile";
 import QuestionFeedList from "../../components/QuestionFeedList/QuestionFeedList";
 import FloatingButton from "../../components/FloatingButton/FloatingButton";
 import QuestionModal from "../../components/QuestionModal/QuestionModal";
-import getSubjectInfo from "../../utils/postpageAPI/getSubjectInfo";
-import getSubjectQuestion from "../../utils/postpageAPI/getSubjectQuestion";
+import { useUserProfileAndQuestions } from "../../hooks/useUserProfileAndQuestions";
 
 export default function PostPage() {
   const { id } = useParams(); // Access the id from route parameters
   const [showModal, setShowModal] = useState(false);
   const [buttonText, setButtonText] = useState("질문하러 가기");
-  const [userProfile, setUserProfile] = useState({});
-  const [userQuestions, setUserQuestions] = useState([]);
-
-  const fetchDataAndSetUserProfile = useCallback(async () => {
-    try {
-      let result = await getSubjectInfo(id);
-      setUserProfile(result);
-    } catch (error) {
-      console.error("피드 페이지를 가져오는 중에 오류가 발생했습니다:", error);
-    }
-  }, [id]);
-
-  const fetchDataAndSetUserQuestions = useCallback(async () => {
-    try {
-      let result = await getSubjectQuestion(id);
-      setUserQuestions(result);
-    } catch (error) {
-      console.error("질문 목록을 가져오는 중에 오류가 발생했습니다:", error);
-    }
-  }, [id]);
-
-  const handleLoad = useCallback(async () => {
-    await fetchDataAndSetUserProfile(id);
-    await fetchDataAndSetUserQuestions(id);
-  }, [fetchDataAndSetUserProfile, fetchDataAndSetUserQuestions, id]);
+  const { userProfile, userQuestions } = useUserProfileAndQuestions(id);
 
   const handleModalOpen = () => {
     setShowModal(true);
@@ -55,12 +30,12 @@ export default function PostPage() {
         setButtonText("질문하러 가기");
       }
     }
-    handleLoad();
+
     window.addEventListener("resize", handleResize);
     handleResize();
 
     return () => window.removeEventListener("resize", handleResize);
-  }, [handleLoad]);
+  }, []);
 
   return (
     <div className={styles.container}>
